backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
Pillow>=10.0.0
xhtml2pdf>=0.2.11
djangorestframework-simplejwt>=5.3.0
